Rename TropoDB persistence files
