+++ implementation/rocksdb/db/zns_impl/impl/tropodb_impl.cc
@@ -23,9 +23,9 @@
 #include "db/zns_impl/index/tropodb_version.h"
 #include "db/zns_impl/index/tropodb_version_set.h"
 #include "db/zns_impl/io/szd_port.h"
-#include "db/zns_impl/persistence/zns_manifest.h"
-#include "db/zns_impl/persistence/zns_wal.h"
-#include "db/zns_impl/persistence/zns_wal_manager.h"
+#include "db/zns_impl/persistence/tropodb_manifest.h"
+#include "db/zns_impl/persistence/tropodb_wal.h"
+#include "db/zns_impl/persistence/tropodb_wal_manager.h"
 #include "db/zns_impl/table/zns_sstable_manager.h"
 #include "db/zns_impl/table/zns_table_cache.h"
 #include "db/zns_impl/utils/tropodb_logger.h"
