+++ implementation/rocksdb/db/zns_impl/index/tropodb_version_edit.h
@@ -9,7 +9,7 @@
 #include "db/dbformat.h"
 #include "db/zns_impl/index/tropodb_version.h"
 #include "db/zns_impl/io/szd_port.h"
-#include "db/zns_impl/persistence/zns_manifest.h"
+#include "db/zns_impl/persistence/tropodb_manifest.h"
 #include "db/zns_impl/ref_counter.h"
 #include "db/zns_impl/table/zns_sstable_manager.h"
 #include "db/zns_impl/table/zns_zonemetadata.h"
