+++ implementation/rocksdb/db/zns_impl/persistence/tropodb_manifest.h
@@ -5,7 +5,7 @@
 
 #include "db/zns_impl/utils/tropodb_diagnostics.h"
 #include "db/zns_impl/io/szd_port.h"
-#include "db/zns_impl/persistence/zns_committer.h"
+#include "db/zns_impl/persistence/tropodb_committer.h"
 #include "db/zns_impl/ref_counter.h"
 #include "rocksdb/slice.h"
 #include "rocksdb/status.h"
