+++ implementation/rocksdb/db/zns_impl/persistence/tropodb_wal_manager.ipp
@@ -10,9 +10,9 @@
 #include "db/zns_impl/tropodb_config.h"
 #include "db/zns_impl/io/szd_port.h"
 #include "db/zns_impl/memtable/tropodb_memtable.h"
-#include "db/zns_impl/persistence/zns_committer.h"
-#include "db/zns_impl/persistence/zns_wal.h"
-#include "db/zns_impl/persistence/zns_wal_manager.h"
+#include "db/zns_impl/persistence/tropodb_committer.h"
+#include "db/zns_impl/persistence/tropodb_wal.h"
+#include "db/zns_impl/persistence/tropodb_wal_manager.h"
 #include "db/zns_impl/utils/tropodb_logger.h"
 #include "port/port.h"
 #include "rocksdb/slice.h"
